+++ cwepr/plotting.py
@@ -31,6 +31,17 @@ connection between a single dataset and such a plot in sense of
 representations stored in datasets. Plotters handling multiple datasets should
 always inherit from the :class:`aspecd.plotting.MultiPlotter` class.
 
+In a certain sense, there is a third type of plotters:
+
+* Plotters consisting of more than one axes
+
+  Shall be derived from :class:`aspecd.plotting.CompositePlotter`.
+
+However, practically mostly these composite plotters will behave like
+plotters handling either single or multiple datasets. Generally,
+these composite plotters will use other types of plotters to perform the
+actual plot tasks. This modular approach allows for great flexibility.
+
 
 A note on array dimensions and axes
 ===================================
@@ -88,6 +99,11 @@ Concrete plotters for single datasets
   Basic 2D plots for single datasets, allowing to plot a series of 2D plots,
   including contour plots and image-type display
 
+* :class:`aspecd.plotting.SinglePlotter2DStacked`
+
+  Stacked plots of 2D data, converting a 2D display into a series of 1D line
+  plots stacked on top of each other.
+
 * :class:`cwepr.plotting.SingleCompositePlotter`
 
   Composite plotter for single datasets, allowing to plot different views of
@@ -111,6 +127,20 @@ Concrete plotters for multiple datasets
 
   Stacked line plots for multiple datasets, allowing to plot a series of
   line-type plots, including (semi)log plots
+
+
+A note for developers
+=====================
+
+As each kind of spectroscopy comes with own needs for extensions, there is a
+class :class:`PlotterExtensions` that can be used as a mixin class for other
+plotters to provide additional functionality for all plotters.
+
+Make sure when implementing functionality here that it really works with all
+types of plotters, *i.e.* both SinglePlotters and MultiPlotters. This is
+particularly relevant if you need to get information from dataset(s),
+as a SinglePlotter will have an attribute ``dataset``, while a MultiPlotter
+will have an attribute ``datasets``.
 
 
 Module documentation
@@ -240,7 +270,61 @@ class GoniometerSweepPlotter(aspecd.plotting.SingleCompositePlotter):
         self.plotter[2].parameters['ytickcount'] = 19
 
 
-class SinglePlotter1D(aspecd.plotting.SinglePlotter1D):
+class PlotterExtensions:
+    """Extensions for plots of cw-EPR data.
+
+    This class is meant as a mixin class for plotters of the cwepr package
+    and provides functionality specific for cw-EPR-spectroscopic data.
+
+    Hence it can only be used as mixin in addition to a plotter class.
+
+    Attributes
+    ----------
+    parameters : :class:`dict`
+        All parameters necessary for the plot, implicit and explicit
+
+        The following keys exist, in addition to those defined by the actual
+        plotter:
+
+        g-axis: :class:`bool`
+            Whether to show an additional *g* axis opposite of the magnetic
+            field axis
+
+            This assumes the magnetic field axis to be the *x* axis and the
+            magnetic field unit to be millitesla (mT).
+
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.parameters['g-axis'] = False
+
+    def _create_g_axis(self, mw_freq=None):
+        """
+        Add a *g* axis as second axis opposite the magnetic field axis.
+
+        Currently, this function assumes the magnetic field axis to be the
+        *x* axis. Additionally, the magnetic field values are assumed to be
+        in millitesla (mT), and the microwave frequency to be in gigahertz (
+        GHz).
+
+        Parameters
+        ----------
+        mw_freq : :class:`float`
+            microwave frequency (**in GHz**) used to convert from mT to g
+
+        """
+        def forward(values):
+            return utils.convert_mT2g(values, mw_freq=mw_freq)
+
+        def backward(values):
+            return utils.convert_g2mT(values, mw_freq=mw_freq)
+
+        gaxis = self.ax.secondary_xaxis('top', functions=(backward, forward))
+        gaxis.set_xlabel('$g\ value$')
+
+
+class SinglePlotter1D(aspecd.plotting.SinglePlotter1D, PlotterExtensions):
     """1D plots of single datasets.
 
     Convenience class taking care of 1D plots of single datasets.
@@ -249,19 +333,8 @@ class SinglePlotter1D(aspecd.plotting.SinglePlotter1D):
     ASpecD documentation of the :class:`aspecd.plotting.SinglePlotter1D`
     class for details.
 
-
-    Attributes
-    ----------
-    parameters : :class:`dict`
-        All parameters necessary for the plot, implicit and explicit
-
-        The following keys exist, in addition to those of the superclass:
-
-        g-axis: :class:`bool`
-            Whether to show an additional opposite of the magnetic field axis
-
-            This assumes the magnetic field axis to be the *x* axis and the
-            magnetic field unit to be millitesla (mT).
+    Furthermore, the class inhertis all functionality from
+    :class:`PlotterExtensions`. See there for additional details.
 
 
     Examples
@@ -282,7 +355,8 @@ class SinglePlotter1D(aspecd.plotting.SinglePlotter1D):
            filename: output.pdf
 
 
-    In case you would have a *g* axis plotted as a second *x* axis on top:
+    In case you would like to have a *g* axis plotted as a second *x* axis on
+    top:
 
     .. code-block:: yaml
 
@@ -295,29 +369,13 @@ class SinglePlotter1D(aspecd.plotting.SinglePlotter1D):
 
     """
 
-    def __init__(self):
-        super().__init__()
-        self.parameters['g-axis'] = False
-
     def _create_plot(self):
         super()._create_plot()
         if self.parameters['g-axis'] and self.dataset.data.axes[0].unit == 'mT':
-            self._create_g_axis()
-
-    def _create_g_axis(self):
-        mw_freq = self.dataset.metadata.bridge.mw_frequency.value
-
-        def forward(values):
-            return utils.convert_mT2g(values, mw_freq=mw_freq)
-
-        def backward(values):
-            return utils.convert_g2mT(values, mw_freq=mw_freq)
-
-        gaxis = self.ax.secondary_xaxis('top', functions=(backward, forward))
-        gaxis.set_xlabel('$g\ value$')
+            self._create_g_axis(self.dataset.metadata.bridge.mw_frequency.value)
 
 
-class SinglePlotter2D(aspecd.plotting.SinglePlotter2D):
+class SinglePlotter2D(aspecd.plotting.SinglePlotter2D, PlotterExtensions):
     """2D plots of single datasets.
 
     Convenience class taking care of 2D plots of single datasets.
@@ -325,6 +383,10 @@ class SinglePlotter2D(aspecd.plotting.SinglePlotter2D):
     As the class is fully inherited from ASpecD for simple usage, see the
     ASpecD documentation of the :class:`aspecd.plotting.SinglePlotter2D`
     class for details.
+
+    Furthermore, the class inhertis all functionality from
+    :class:`PlotterExtensions`. See there for additional details.
+
 
     Examples
     --------
@@ -391,10 +453,111 @@ class SinglePlotter2D(aspecd.plotting.SinglePlotter2D):
     Make sure to check the documentation of the ASpecD
     :mod:`aspecd.plotting` module for further parameters that can be set.
 
+    In case you would like to have a *g* axis plotted as a second *x* axis on
+    top:
+
+    .. code-block:: yaml
+
+       - kind: singleplot
+         type: SinglePlotter2D
+         properties:
+           parameters:
+             g-axis: true
+           filename: output.pdf
+
     """
 
+    def _create_plot(self):
+        super()._create_plot()
+        if self.parameters['g-axis'] and self.dataset.data.axes[0].unit == 'mT':
+            self._create_g_axis(self.dataset.metadata.bridge.mw_frequency.value)
 
-class MultiPlotter1D(aspecd.plotting.MultiPlotter1D):
+
+class SinglePlotter2DStacked(aspecd.plotting.SinglePlotter2DStacked,
+                             PlotterExtensions):
+    """Stacked plots of 2D data
+
+    A stackplot creates a series of lines stacked on top of each other from
+    a 2D dataset.
+
+    As the class is fully inherited from ASpecD for simple usage, see the
+    ASpecD documentation of the :class:`aspecd.plotting.SinglePlotter2DStacked`
+    class for details.
+
+    Furthermore, the class inhertis all functionality from
+    :class:`PlotterExtensions`. See there for additional details.
+
+
+    Examples
+    --------
+    For convenience, a series of examples in recipe style (for details of
+    the recipe-driven data analysis, see :mod:`aspecd.tasks`) is given below
+    for how to make use of this class. Of course, all parameters settable
+    for the superclasses can be set as well. The examples focus each on a
+    single aspect.
+
+    In the simplest case, just invoke the plotter with default values:
+
+    .. code-block:: yaml
+
+       - kind: singleplot
+         type: SinglePlotter2DStacked
+         properties:
+           filename: output.pdf
+
+    If you need to more precisely control the formatting of the y tick
+    labels, particularly the number of decimals shown, you can set the
+    formatting accordingly:
+
+    .. code-block:: yaml
+
+       - kind: singleplot
+         type: SinglePlotter2DStacked
+         properties:
+           filename: output.pdf
+           parameters:
+             yticklabelformat: '%.2f'
+
+    In this particular case, the y tick labels will appear with only two
+    decimals. Note that currently, the "old style" formatting specifications
+    are used due to their widespread use in other programming languages and
+    hence the familiarity of many users with this particular notation.
+
+    Sometimes you want to have horizontal "zero lines" appear for each
+    individual trace of the stacked plot. This can be achieved explicitly
+    setting the "show_zero_lines" parameter to "True" that is set to "False"
+    by default:
+
+    .. code-block:: yaml
+
+       - kind: singleplot
+         type: SinglePlotter2DStacked
+         properties:
+           filename: output.pdf
+           parameters:
+             show_zero_lines: True
+
+    In case you would like to have a *g* axis plotted as a second *x* axis on
+    top:
+
+    .. code-block:: yaml
+
+       - kind: singleplot
+         type: SinglePlotter2DStacked
+         properties:
+           parameters:
+             g-axis: true
+           filename: output.pdf
+
+    """
+
+    def _create_plot(self):
+        super()._create_plot()
+        if self.parameters['g-axis'] and self.dataset.data.axes[0].unit == 'mT':
+            self._create_g_axis(self.dataset.metadata.bridge.mw_frequency.value)
+
+
+class MultiPlotter1D(aspecd.plotting.MultiPlotter1D, PlotterExtensions):
     """1D plots of multiple datasets.
 
     Convenience class taking care of 1D plots of multiple datasets.
@@ -402,6 +565,10 @@ class MultiPlotter1D(aspecd.plotting.MultiPlotter1D):
     As the class is fully inherited from ASpecD for simple usage, see the
     ASpecD documentation of the :class:`aspecd.plotting.MultiPlotter1D`
     class for details.
+
+    Furthermore, the class inhertis all functionality from
+    :class:`PlotterExtensions`. See there for additional details.
+
 
     Examples
     --------
@@ -444,10 +611,30 @@ class MultiPlotter1D(aspecd.plotting.MultiPlotter1D):
         ``#``, you need to explicitly tell YAML that these are strings,
         surrounding the values by quotation marks.
 
+    In case you would like to have a *g* axis plotted as a second *x* axis on
+    top:
+
+    .. code-block:: yaml
+
+       - kind: multiplot
+         type: MultiPlotter1D
+         properties:
+           parameters:
+             g-axis: true
+           filename: output.pdf
+
     """
 
+    def _create_plot(self):
+        super()._create_plot()
+        if self.parameters['g-axis'] \
+                and self.datasets[0].data.axes[0].unit == 'mT':
+            self._create_g_axis(
+                self.datasets[0].metadata.bridge.mw_frequency.value)
 
-class MultiPlotter1DStacked(aspecd.plotting.MultiPlotter1DStacked):
+
+class MultiPlotter1DStacked(aspecd.plotting.MultiPlotter1DStacked,
+                            PlotterExtensions):
     """Stacked 1D plots of multiple datasets.
 
     Convenience class taking care of 1D plots of multiple datasets.
@@ -455,6 +642,9 @@ class MultiPlotter1DStacked(aspecd.plotting.MultiPlotter1DStacked):
     As the class is fully inherited from ASpecD for simple usage, see the
     ASpecD documentation of the :class:`aspecd.plotting.MultiPlotter1DStacked`
     class for details.
+
+    Furthermore, the class inhertis all functionality from
+    :class:`PlotterExtensions`. See there for additional details.
 
     Examples
     --------
@@ -511,50 +701,23 @@ class MultiPlotter1DStacked(aspecd.plotting.MultiPlotter1DStacked):
            parameters:
              show_zero_lines: True
 
-    """
+    In case you would like to have a *g* axis plotted as a second *x* axis on
+    top:
 
+    .. code-block:: yaml
 
-class CompositePlotter(aspecd.plotting.CompositePlotter):
-    """Base class for plots consisting of multiple axes.
-
-    The underlying idea of composite plotters is to use a dedicated
-    existing plotter for each axis and assign this plotter to the list of
-    plotters of the CompositePlotter object. Thus the actual plotting task
-    is left to the individual plotter and the CompositePlotter only takes
-    care of the specifics of plots consisting of more than one axis.
-
-    As the class is fully inherited from ASpecD for simple usage, see the
-    ASpecD documentation of the :class:`aspecd.plotting.CompositePlotter`
-    class for details.
+       - kind: multiplot
+         type: MultiPlotter1DStacked
+         properties:
+           parameters:
+             g-axis: true
+           filename: output.pdf
 
     """
 
-
-class SingleCompositePlotter(aspecd.plotting.SingleCompositePlotter):
-    """Composite plotter for single datasets.
-
-    This composite plotter is used for different representations of one and the
-    same dataset in multiple axes contained in one figure. In this respect,
-    it works like all the other ordinary single plotters derived from
-    :class:`SinglePlotter`, *i.e.* it usually gets called by using the dataset's
-    :meth:`aspecd.dataset.Dataset.plot` method.
-
-    As the class is fully inherited from ASpecD for simple usage, see the
-    ASpecD documentation of the :class:`aspecd.plotting.SingleCompositePlotter`
-    class for details.
-
-    """
-
-
-class Saver(aspecd.plotting.Saver):
-    """Base class for saving plots.
-
-    For basic saving of plots, no subclassing is necessary, as the
-    :meth:`save` method uses :meth:`matplotlib.figure.Figure.savefig` and
-    can cope with all possible parameters via the :attr:`parameters` property.
-
-    As the class is fully inherited from ASpecD for simple usage, see the
-    ASpecD documentation of the :class:`aspecd.plotting.Saver` class for
-    details.
-
-    """
+    def _create_plot(self):
+        super()._create_plot()
+        if self.parameters['g-axis'] \
+                and self.datasets[0].data.axes[0].unit == 'mT':
+            self._create_g_axis(
+                self.datasets[0].metadata.bridge.mw_frequency.value)